crypto

+++ src/components/trade.tsx
@@ -27,7 +27,7 @@ const Trade = (props: Props) => {
   
       if (command === "buy" || command === "sell") {
         if (props.symbol.price) {
-          const size = parseInt(args[0].replace('k', '000'))
+          const size = parseFloat(args[0].replace('k', '000'))
           if (!isNaN(size)) {
             positions.addPosition(command === "buy" ? "Buy" : "Sell", size, props.symbol)
           }
